SGS upgrade bugfixes from pre-commit validation
git-svn-id: http://svn.sgscript.org/tacstrike@217 89f9ca44-2ff2-4b64-a49d-2a4f1987ee72

+++ ext/src/sgscript/sgscript.h
@@ -861,7 +861,7 @@ static SGS_INLINE SGSRESULT sgs_StdScriptFSFunc( void* ctx, SGS_CTX, int op, sgs
 				return SGS_EINPROC;
 			}
 			data->userhandle = f;
-			data->size = sz;
+			data->size = (size_t) sz;
 			return SGS_SUCCESS;
 		}
 	case SGS_SFS_FILE_READ:
